reportes/Enlaces Ericsson Tigo/requirements.txt
firebase-admin>=6.0.0
reportlab>=3.6.0
Pillow>=9.0.0
requests>=2.28.0
google-cloud-firestore>=2.10.0


